+++ accounts-line/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## 1.0.1 - 2020-01-01
+* Fix imports
+* Update dependencies
+* Update minimum Meteor version to 1.8.3
+
 ## 1.0.0 - 2018-04-19
 * Initial release