+++ line-config-ui/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## 1.0.1 - 2020-01-01
+* Update dependencies
+* Update minimum Meteor version to 1.8.3
+* Add import statements for templates
+
 ## 1.0.0 - 2018-04-19
 * Initial release